feat: set table to page width
