+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table.dart
@@ -5,6 +5,4 @@ export 'simple_table_more_action.dart';
 export 'simple_table_operations/simple_table_operations.dart';
 export 'simple_table_row_block_component.dart';
 export 'simple_table_shortcuts/simple_table_commands.dart';
-export 'simple_table_widgets/simple_table_add_column_and_row_button.dart';
-export 'simple_table_widgets/simple_table_add_column_button.dart';
-export 'simple_table_widgets/simple_table_add_row_button.dart';
+export 'simple_table_widgets/widgets.dart';
