+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table_widgets/simple_table_basic_button.dart
@@ -1,0 +1,50 @@
+import 'package:appflowy/generated/flowy_svgs.g.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/plugins.dart';
+import 'package:flowy_infra_ui/flowy_infra_ui.dart';
+import 'package:flutter/material.dart';
+
+class SimpleTableBasicButton extends StatelessWidget {
+  const SimpleTableBasicButton({
+    super.key,
+    required this.text,
+    required this.onTap,
+    this.leftIconSvg,
+    this.leftIconBuilder,
+    this.rightIcon,
+  });
+
+  final FlowySvgData? leftIconSvg;
+  final String text;
+  final VoidCallback onTap;
+  final Widget Function(bool onHover)? leftIconBuilder;
+  final Widget? rightIcon;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: SimpleTableConstants.moreActionHeight,
+      padding: SimpleTableConstants.moreActionPadding,
+      child: FlowyIconTextButton(
+        margin: SimpleTableConstants.moreActionHorizontalMargin,
+        leftIconBuilder: _buildLeftIcon,
+        iconPadding: 10.0,
+        textBuilder: (onHover) => FlowyText.regular(
+          text,
+          fontSize: 14.0,
+          figmaLineHeight: 18.0,
+        ),
+        onTap: onTap,
+        rightIconBuilder: (onHover) => rightIcon ?? const SizedBox.shrink(),
+      ),
+    );
+  }
+
+  Widget _buildLeftIcon(bool onHover) {
+    if (leftIconBuilder != null) {
+      return leftIconBuilder!(onHover);
+    }
+    return leftIconSvg != null
+        ? FlowySvg(leftIconSvg!)
+        : const SizedBox.shrink();
+  }
+}
