+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table_widgets/simple_table_divider.dart
@@ -1,0 +1,28 @@
+import 'package:appflowy/plugins/document/presentation/editor_plugins/plugins.dart';
+import 'package:flutter/material.dart';
+
+class SimpleTableRowDivider extends StatelessWidget {
+  const SimpleTableRowDivider({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return VerticalDivider(
+      color: context.simpleTableBorderColor,
+      width: 1.0,
+    );
+  }
+}
+
+class SimpleTableColumnDivider extends StatelessWidget {
+  const SimpleTableColumnDivider({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Divider(
+      color: context.simpleTableBorderColor,
+      height: 1.0,
+    );
+  }
+}
